+++ include/elio/elio.hpp
@@ -17,6 +17,7 @@
 #include "coro/task_execution_context.hpp"
 #include "coro/promise_base.hpp"
 #include "coro/task.hpp"
+#include "coro/task_group.hpp"
 #include "coro/awaitable_base.hpp"
 #include "coro/frame.hpp"
 #include "coro/cancel_token.hpp"